create findAll, deleteAll and getAverageAge methods

+++ task/src/main/java/com/bitergo/task/person/PersonRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface PersonRepository extends CrudRepository<Person,Long> {
     @Query("SELECT AVG(age) FROM Person")
-    int getAverageAge();
+    float getAverageAge();
 }
